base: Introduce Accidental Touch
Accidental Touch is a feature useful for many, especially the gamers out there.
It prevents any sort of accidental touch on hardware buttons while the touchscreen is
in use.
This means you won’t experience any unexpected behaviour while playing games,
web browsing or even when attempting to reach your finger out to the edge of
that humongous display.

Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Change-Id: Ie7074a2cba48ae56a7a0a2bb90ef79f5b84403cf
Signed-off-by: Anushek Prasal <anushekprasal@gmail.com>